--- .Net/Master C# and SQL by Building Applications/CustomerOrderViewer/CustomerOrderViewer.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.4.33110.190
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CustomerOrderViewer", "CustomerOrderViewer\CustomerOrderViewer.csproj", "{D90F168B-9FA2-4264-BD21-046B0B392EF6}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{D90F168B-9FA2-4264-BD21-046B0B392EF6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{D90F168B-9FA2-4264-BD21-046B0B392EF6}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{D90F168B-9FA2-4264-BD21-046B0B392EF6}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{D90F168B-9FA2-4264-BD21-046B0B392EF6}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {587EA042-87EA-44A5-96E6-96B6B4B55EEC}
-	EndGlobalSection
-EndGlobal
