--- .Net/Master C# and SQL by Building Applications/CustomerOrderViewer/CustomerOrderViewer/Repository/CustomerOrderDetailCommand.cs
@@ -1,58 +1,0 @@
-﻿using CustomerOrderViewer.Models;
-using System;
-using System.Collections.Generic;
-using System.Data.SqlClient;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace CustomerOrderViewer.Repository
-{
-    internal class CustomerOrderDetailCommand
-    {
-        private string _connectionString;
-
-        public CustomerOrderDetailCommand(string connectionString)
-        {
-            _connectionString = connectionString;
-        }
-        
-        public IList<CustomerOrderDetailModel> GetList()
-        {
-            List<CustomerOrderDetailModel> customerOrderDetailModels = new List<CustomerOrderDetailModel>();
-
-            using (SqlConnection connection = new SqlConnection(_connectionString))
-            {
-                connection.Open();
-
-                using (SqlCommand command = new SqlCommand("SELECT CustomerOrderId, CustomerId, ItemId, FirstName, LastName,[Description], Price FROM CustomerOrderDetail", connection))
-                {
-                    using(SqlDataReader reader = command.ExecuteReader())
-                    {
-                        if (reader.HasRows)
-                        {
-                            while (reader.Read())
-                            {
-                                CustomerOrderDetailModel customerOrderDetailModel = new CustomerOrderDetailModel()
-                                {
-                                    CustomerOrderId = Convert.ToInt32(reader["CustomerOrderId"]),
-                                    CustomerId = Convert.ToInt32(reader["CustomerId"]),
-                                    ItemId = Convert.ToInt32(reader["ItemId"]),
-                                    FirstName = Convert.ToString(reader["FirstName"]),
-                                    LastName = Convert.ToString(reader["LastName"]),
-                                    Description = Convert.ToString(reader["Description"]),
-                                    Price = Convert.ToDecimal(reader["Price"]),
-                                };
-
-                                customerOrderDetailModels.Add(customerOrderDetailModel);
-
-                            }
-                        }
-                    }
-                }
-            }
-
-                return customerOrderDetailModels;
-        }
-    }
-}
