--- .Net/Master C# and SQL by Building Applications/CustomerOrderViewer2.0/CustomerOrderViewer2.0.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.4.33110.190
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CustomerOrderViewer2.0", "CustomerOrderViewer2.0\CustomerOrderViewer2.0.csproj", "{38517FFE-6B96-44B0-96BC-C034BD445D04}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{38517FFE-6B96-44B0-96BC-C034BD445D04}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{38517FFE-6B96-44B0-96BC-C034BD445D04}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{38517FFE-6B96-44B0-96BC-C034BD445D04}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{38517FFE-6B96-44B0-96BC-C034BD445D04}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {B1258561-7038-486A-AE77-58A97C3C3F0E}
-	EndGlobalSection
-EndGlobal
