--- .Net/Master C# and SQL by Building Applications/CustomerOrderViewer2.0/CustomerOrderViewer2.0/Models/CustomerModel.cs
@@ -1,19 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Security.Cryptography.X509Certificates;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace CustomerOrderViewer2._0.Models
-{
-    internal class CustomerModel
-    {
-        public int CustomerId { get; set; }
-        public string FirstName { get; set; }
-        public string MiddleName { get; set; }
-        public string LastName { get; set; }
-        public int Age { get; set; }
-
-    }
-}
