--- .Net/Master C# and SQL by Building Applications/CustomerOrderViewer2.0/CustomerOrderViewer2.0/Models/CustomerOrderDetailModel.cs
@@ -1,19 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace CustomerOrderViewer2._0.Models
-{
-    internal class CustomerOrderDetailModel
-    {
-        public int CustomerOrderId { get; set; }
-        public int CustomerId { get; set; }
-        public int ItemID { get; set; }
-        public string FirstName { get; set; }
-        public string LastName { get; set; }
-        public string Description { get; set; }
-        public decimal Price { get; set; }
-    }
-}
