--- .Net/Master C# and SQL by Building Applications/CustomerOrderViewer2.0/CustomerOrderViewer2.0/Models/ItemModel.cs
@@ -1,15 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace CustomerOrderViewer2._0.Models
-{
-    internal class ItemModel
-    {
-        public int ItemId { get; set; }
-        public string Description { get; set; }
-        public decimal Price { get; set; }
-    }
-}
