--- .Net/Master C# and SQL by Building Applications/CustomerOrderViewer2.0/README.md
@@ -1,8 +1,0 @@
-# Customer Order Viewer v2
-
-Based on O'Reilly Course "Master C# and SQL by Building Applications" by Avetis Ghukasyan <br>
-https://learning.oreilly.com/videos/master-c-and/9781800560093/<br><br>
-
-Added Update and Store Procedures for Customer as well as Item.<br>
-Also Changed Console App menu, broken down by action; Retrieve, Update, and Delete.<br>
-Then broke down to submenus for different categories; Orders, Customers, and Items.
