--- .Net/Master C# and SQL by Building Applications/README.md
@@ -1,4 +1,0 @@
-# Master C# and SQL by Building Applications
-
-Based on O'Reilly Course "Master C# and SQL by Building Applications" by Avetis Ghukasyan <br>
-https://learning.oreilly.com/videos/master-c-and/9781800560093/<br><br>
