--- .Net/SimpleWebScraper/README.md
@@ -1,5 +1,0 @@
-# Simple Web Scraper
-
-Based on O'Reilly Course "Learn C# by Building Applications" by Avetis Ghukasyan <br>
-https://learning.oreilly.com/videos/learn-c-by/9781800562721/<br><br>
-Scrapes Craigslist, but propably would change to have it scrape something more usful.
