--- .Net/SimpleWebScraper/SimpleWebScraper.Test.Unit/Usings.cs
@@ -1,1 +1,0 @@
-global using Microsoft.VisualStudio.TestTools.UnitTesting;
