--- .Net/SimpleWebScraper/SimpleWebScraper.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.4.33110.190
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SimpleWebScraper", "SimpleWebScraper\SimpleWebScraper.csproj", "{7C561B27-3EBE-48C2-B696-49EE37177CD7}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SimpleWebScraper.Test.Unit", "SimpleWebScraper.Test.Unit\SimpleWebScraper.Test.Unit.csproj", "{D3AE1649-B2A5-40CB-B73A-D3A633A525FB}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{7C561B27-3EBE-48C2-B696-49EE37177CD7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{7C561B27-3EBE-48C2-B696-49EE37177CD7}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{7C561B27-3EBE-48C2-B696-49EE37177CD7}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{7C561B27-3EBE-48C2-B696-49EE37177CD7}.Release|Any CPU.Build.0 = Release|Any CPU
-		{D3AE1649-B2A5-40CB-B73A-D3A633A525FB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{D3AE1649-B2A5-40CB-B73A-D3A633A525FB}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{D3AE1649-B2A5-40CB-B73A-D3A633A525FB}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{D3AE1649-B2A5-40CB-B73A-D3A633A525FB}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {1C43893E-27CE-4B16-93A1-A98E499458B8}
-	EndGlobalSection
-EndGlobal
