--- .Net/SimpleWebScraper/SimpleWebScraper/Builders/ScrapeCriteriaPartBuilder.cs
@@ -1,44 +1,0 @@
-﻿using SimpleWebScraper.Data;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Text.RegularExpressions;
-using System.Threading.Tasks;
-
-namespace SimpleWebScraper.Builders
-{
-    internal class ScrapeCriteriaPartBuilder
-    {
-        private string _regex;
-        private RegexOptions _regexOption;
-
-        public ScrapeCriteriaPartBuilder()
-        {
-            SetDefaults();
-        }
-
-        private void SetDefaults()
-        {
-            _regex = string.Empty;
-            _regexOption = RegexOptions.None;
-        }
-        public ScrapeCriteriaPartBuilder WithRegex(string regex)
-        {
-            _regex = regex;
-            return this;
-        }
-        public ScrapeCriteriaPartBuilder WithRegexOption(RegexOptions regexOption)
-        {
-            _regexOption = regexOption;
-            return this;
-        }
-        public ScrapeCriteriaPart Build()
-        {
-            ScrapeCriteriaPart scrapeCriteriaPart = new ScrapeCriteriaPart();
-            scrapeCriteriaPart.Regex = _regex;
-            scrapeCriteriaPart.RegexOption = _regexOption;
-            return scrapeCriteriaPart;
-        }
-    }
-}
