--- .Net/SimpleWebScraper/SimpleWebScraper/Data/ScrapeCriteria.cs
@@ -1,21 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Text.RegularExpressions;
-using System.Threading.Tasks;
-
-namespace SimpleWebScraper.Data
-{
-    internal class ScrapeCriteria
-    {
-        public ScrapeCriteria()
-        {
-            Parts = new List<ScrapeCriteriaPart>();
-        }
-        public string Data { get; set; }
-        public string Regex { get; set; }
-        public RegexOptions RegexOption { get; set; }
-        public List<ScrapeCriteriaPart> Parts { get; set; }
-    }
-}
