--- .Net/SimpleWebScraper/SimpleWebScraper/Data/ScrapeCriteriaPart.cs
@@ -1,15 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Text.RegularExpressions;
-using System.Threading.Tasks;
-
-namespace SimpleWebScraper.Data
-{
-    internal class ScrapeCriteriaPart
-    {
-        public string Regex { get; set; }
-        public RegexOptions RegexOption { get; set; }
-    }
-}
