--- .Net/SimpleWebScraper/SimpleWebScraper/SimpleWebScraper.csproj
@@ -1,15 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-	<PropertyGroup>
-		<OutputType>Exe</OutputType>
-		<TargetFramework>net7.0</TargetFramework>
-		<ImplicitUsings>enable</ImplicitUsings>
-		<Nullable>enable</Nullable>
-	</PropertyGroup>
-
-    <ItemGroup>
-	    <AssemblyAttribute Include="System.Runtime.CompilerServices.InternalsVisibleToAttribute">
-		<_Parameter1>SimpleWebScraper.Test.Unit</_Parameter1>
-	    </AssemblyAttribute>
-    </ItemGroup>
-</Project>
